requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
ffmpeg-python>=0.2.0
pyannote.audio==2.1.1
transformers>=4.27.0
scipy>=1.10.0
omegaconf>=2.3.0
protobuf>=3.20.0
torchmetrics>=0.11.4 